taskmanager-backend/requirements.txt
Django
djangorestframework
psycopg2-binary
celery
django-celery-beat
redis
ulid-py
django-cors-headers
requests
aiogram
aiogram_dialog
aiohttp